[refactor] split all endpoints
